+++ src/tools.ts
@@ -1,0 +1,30 @@
+import * as fs from 'fs';
+import * as path from 'path';
+
+export function mkdirs(dirname, callback) {
+    fs.exists(dirname, function (exists) {
+        if (exists) {
+            callback();
+        } else {
+            mkdirs(path.dirname(dirname), function () {
+                fs.mkdir(dirname, callback);
+            });
+        }
+    });
+}
+
+export function mkdirsSync(dirname) {
+    if (fs.existsSync(dirname)) {
+        return true;
+    } else {
+        if (mkdirsSync(path.dirname(dirname))) {
+            fs.mkdirSync(dirname);
+            return true;
+        }
+    }
+}
+
+export function isSubPath(from: string, to: string): boolean {
+    let rel = path.relative(to, from);
+    return !(path.isAbsolute(rel) || rel.substr(0, 2) == "..")
+}